Validate fn "name" when enforcing strict mode
TC39 reached consensus in March 2018 to extend the definition of the
term "function code" to include the BindingIdentifier informally
referred to as the function's "name" in those productions which include
it [1]. This has the effect of restricting the set of valid identifiers
for functions whose bodies include a "use strict" directive.

Because the project's test suite does not include tests for this case,
no change to existing test material is necessary. However, the
`test262-parser-tests` project *does* require modification, so this
patch should not be applied until that project has been updated
accordingly [2].

[1] https://github.com/tc39/ecma262/pull/1158
[2] https://github.com/tc39/test262-parser-tests/pull/17