Test262 (#395)
This adds test262 as a dev dependency, and asserts that every test in it parses or fails to parse as appropriate.

The bulk of this PR is test expectations. These are split into files for ease of maintenance.

Managing expectations is a lot easier since the project adopted feature flags. Part of the test expectations is a list of flags we expect to indicate syntax which isn't parseable; some of these are parseable despite such a flag, and need to be explicitly whitelisted.

The test runner enforces that expectations are not stale: every named test must exist, and every test which is whitelisted as in the above paragraph must refer to an xfailed feature.

No tests are skipped, nor is a mechanism provided to do so.

Fixes #362.